requirements.txt
aiogram
peewee
python-decouple
psycopg2-binary